Refactor CharcoalConfig and UIKitConfigs
